requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.58.0
django-appconf==1.0.5
django-compressor==4.4
django-crispy-forms==2.1
django-extensions==3.2.3
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.14.0
frozenlist==1.4.0
glob2==0.7
httpie==3.2.2
idna==3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.1
oauthlib==3.2.2
pandas==2.1.2
pip-upgrade==0.0.6
pkg_resources==0.0.0
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
rcssmin==1.1.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
twilio==8.10.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==3.0.1
whitenoise==6.6.0
yarl==1.9.2
